Pytest prints out what target is used
To get this working, I had to make a `conftest.py` in the root
directory, as pytest will only look in subdirectories for conftest.py
if those tests are explicitly specified at the command line.
I put the parser option in there as well so that you can specify
the target when running pytest on all tests.