Split CxGroup into more specific groups
The CxGroup was doing some compartment stuff, some synpase stuff,
some axon stuff, and some probe stuff. Now each of those four
groups of things are tracked separately and managed by the
CoreGroup container class. CoreGroup might not be a great name
as I'm not sure how the group interacts with hardware cores, but
it will do for now.