First draft of documentation.
The overall structure of the documentation is copied from other
Nengo projects. Much of the content is adapted from the docs
currently in the wiki, with slight changes to make it more generic
to any Loihi setup and not necessarily ours. The overview
and API docs are new. Some classes that made sense to include in
the API docs had sparse docstrings, so those were added where
appropriate. Examples were moved to the docs folder for easier
inclusion and fewer folders in the root directory.