Disable rectified linear and scale weights
This reverts part of the "Fix dt scaling on connections" commit,
which was partly done in an effort to fix RectifiedLinear neurons,
which have different scaling properties than LIF neurons.
Those issues still exist, but rather than hold up the release
fixing them, we instead disable ReLu neurons for the time being
and add back the weight scaling.

This commit reverts the changes to test_ens_neurons_on_host,
but also modifies the test to use better variable names and
add comments as to what the test is supposed to be doing.