Fix dt scaling fix.
I added a test for ensemble to neuron connections, which failed
because the weights were scaled up by 1/dt. However, the node to
neuron connections were scaled properly, so what ended up working
was to revert the previous reverting of the dt scaling.
However, doing this caused `test_n2n_on_host` to fail when the pre
ensemble was simulated off chip and the data sent into the post
neuron which were on chip, so to fix that the transform when
pre is a `ChipReceiveNeurons` object is now scaled by 1/dt.