Make running Nengo tests easy
To run Nengo tests, do

    pytest --pyargs nengo

and pass any other arguments as you would when running the
Nengo Loihi test suite.

This commit includes changes to better handle different
neuron types. Before, we were only checking that `nengo.Direct`
ensembles are run offchip, but in actuality we want to run
any ensemble that isn't a LIF or SpikingReLu offchip.

This commit also adds more information to various errors and
assertions as they get raised in the Nengo test suite. There
is a notebook in the sandbox that reads these errors messages
to determine why an error occurred in order to group issues
to see which issues are the most pressing. In this commit, these
groups are reflected in the `Simulator.unsupported` list.