loihi_api: Split validation errors for clarity
It will now be clear if a value (e.g. `vth`) is too small
or too large. Before, we lumped these two cases together.
This will help avoid some confusion, like that in #59.

`vth_to_manexp` now checks that `vth > 0` instead of `>= 0`.
This matches the validation done when using Loihi.