Make interneuron parameters configurable
The default interrneuron firing rate defaults to
`1 / (dt * inter_n)` so that the sum of the max rates of all
interneurons adds to the max firing rate of 1/dt. Since this sum
of max rates is used for the interneurons from nodes in
`splitter.py`, this ensures that models can run with any
`dt` without needing to change `inter_rate`.
However, if necessary, the `inter_rate` can still be manually
overridden.

Also made a small reduction in duplication in Simulator.__init__.