Add strict mode to emulator
In strict mode, issues (like U/V overflows) will raise a
SimulatorException. When not in strict mode, these issues will
be raised as a warning. The default is not to be in strict mode,
as issues do happen on occasion, but we run the test suite in
strict mode to find issues early.