Do Node filtering before interneurons
This makes Nodes like Ensembles in this respect, and should avoid
problems with different taus on multiple inputs to an ensemble.

The warning for having no synapse on a connection from a Node is
now impossible to trip since Nodes are always off the chip and
the splitter will always use an `inter_tau` synapse. So I removed
the warning and the associated test.

Fixes #94.