Refactor and test splitter
The main change is the introduction of the SplitNetworks class,
which keeps track of what's going on during the splitting process.
This makes the placing and splitting functions easier to debug
and test, as evidenced by the new tests added to test_splitter.py.

Most of the behavior from the previous splitter is unchanged
in this commit, though some subtle bugs may have been fixed through
the refactoring process. One deliberate improvement is to
explicitly place ensemble acting as the `pre` of a learned
connection off-chip, unless this is overridden by the user (though
this will end up resulting in an exception).