Allow population connections on Loihi
This includes a workaround for compiling population axons to chip.
The NxSDK axon compiler currently does not support population
(pop16, pop32) axons. This commit does the axon compiling ourselves
for these cases. All axons have to be collected and built together,
since each compartment must target a sequential group of axons.

Also, when building CxSpikeInputs, we were adding axons. This is not
required since axons are sent via a BasicSpikeGenerator or snips.
So this commit removes those axons.