Modifying tau_s gives a warning, not an exception
When multiple tau_s values are requested, we use the longest
requested tau_s. That means that if the user requests tau_s=0
we will use it despite it being shorter than the default 0.005,
and if the user requests 0.1 and then 0.05 we will use 0.1
because it is the longest requested tau_s.

Co-authored-by: Eric Hunsberger <erichuns@gmail.com>