Decrease lower bound on default intercept
Since intercepts are applied after the encoders, it's only the upper
bound that needs to be less than one to avoid high gains. The lower
bound should go to -1.

This changed the intercepts for test_nengo_comm_channel_compare,
causing it to fail. The test was fragile because it did not apply
sufficient filtering to the output. The new test filters the outputs
more and can thus have tigher tolerances (passing on 10 seeds).