Refactor interneurons, now called decode neurons
The biggest change is to be able to customize how many decode
neurons are used and how they are organized.
Additionally, the name "interneurons" was changed to "decode
neurons" to avoid confusion with the concept of inhibitory
interneurons in neuroscience.

This commit also removes sharing encoders between node inputs
and ensemble inputs. This greatly simplifies things, because these
no longer have to be identical.

Several decode neuron implementations are provided:
- OnOffDecodeNeurons uses identical pairs of on/off neurons for
  each dimension. This is the default for generating input spikes
  (for nodes).
- NoisyDecodeNeurons uses pairs of on/off neurons for each dimension
  with injected noise so that each on/off pair is different.
  This was the default behavior prior to this commit for decoded
  connections between two ensembles.
- Preset5DecodeNeurons uses five heterogeneous on/off pairs with
  pre-set values for each dimension.
- Preset10DecodeNeurons uses ten heterogeneous on/off pairs with
  pre-set values for each dimension. This is now the default behavior
  for decoded connections between two ensembles.

This commit uncovered a bug in test_communication.test_neurons2node
in which a magic number on the chip was being treated as a spike.
For now, we set `precompute=True` to avoid the bug.

Partially addresses #129.