Refactor host-and-chip communication
The main goal was to keep CxSimulator and LoihiSimulator from
having to know about Nengo objects (i.e. those used in the builder
Model). This allows CxSimulator and LoihiSimulator to work with
general CxModel objects again, not just builder.Model objects with
the `probes` attribute (among other things).

This also adds a new SpikeInput class that handles spike inputs
to Loihi, so that axon information is stored more transparently.