Add support for Nengo convolution transforms
And update the MNIST example with new syntax and parameter files.

This is not backwards compatible: We do not support the new
convolution code with older Nengo, since much of the supporting code
is in new Nengo. However, this is not a significant regression,
since convolution only worked briefly with old Nengo.