Make learning deterministic
In the emulator, emulator/interface.py and passthrough.py now
use OrderedDicts where needed, since order is not guaranteed in
normal dictionaries.

On the chip, we now set the seed for all random objects so that
we have control over the on-chip randomness. This commit includes
a test to check that the same simulator seed results in same learning,
and different simulator seeds result in different learning.