Fix noise for small exponents
The chip allows noise exponents down to zero; this now works on the
emulator as well.

However, a known abnormality in shifting on the chip means that
noise exponents less than 7 act as one greater than they should.
This commit addresses that problem by subtracting one from
weight exponents before putting them on the chip.

We also improve the test for noise, so that it tests different
levels of noise and asserts that the statistics of the resulting
Wiener process are as expected.