Clarify negative cx_base value for ignoring axon
A negative cx_base value is sometimes helpful to indicate an axon
that is not required when making convolutional connections. The chip
itself does not support negative cx_base values, so we just remove
this axon when building the model. We used to use a value of -2048
to indicate this, but this led to the emulator misleadingly allowing
smaller negative values. Now, we just use -1 to indicate an unused
axon, and give an error for other values.

Also added a test of negative cx_base values working as expected.