Combine splitter into builder
Most of the work done by the splitter is now done in the builder.
This should give more clarity and control over the mapping between
pre-build and post-build objects. The `Split` class takes on the
organizational tasks of the old `Splitter`, giving directives to
the builder about what should be on- or off-chip.

Several tests were added or modified to ensure that these changes
did not affect built models in adverse ways. Some tests were added
to ensure that the refactoring made improvements; for example, the
refactored splitter no longer mutates networks, which closes #211.
Simulations should now be identical whether precompute is True
or False, so the tolerance in test_precompute is now set to zero.

The passthrough removal process was also modified to act in the
same way as `Split`, and is now named `PassthroughSplit` accordingly.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>