Revert "Parallelize hardware tests"
This partially reverts commit
70ecb8a475985846c81bc5e8059056042f855a34.

The patch to graph.Graph seems to be incomplete, as one process
could be writing to the file to be copied while another process
is attempting to read it. The includes are also not copied, so
likely we need a more robust solution for writing snip files to
temporary directories.

The command was only changed to request one worker instead of two.