Remove value from PES error clip warning
Warnings normally are only shown once, but since this warning
included the exact value it would show many times, which was
causing a lot of noise in example notebooks.

Since the exact PES error value can be useful for debugging, we
now log the exact value at the DEBUG level instead.

Fixes #171.