Make default allocator multi-chip
The default allocator is now Greedy, which uses all of the cores
on one chip then proceeds to the next chip. Each block is only
allocated to one core.

The previous default, OneToOne, has been removed because it
offers no functionality over the Greedy allocator. The OneToOne
allocator is (mostly) equivalent to Greedy(n_chips=1).

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>