Add debug statements for failed copies
This seems to be the underlying issue when two processes attempt
to build Loihi models at the same time. Reproducing the error and
fully debugging it is time consuming, so we're adding these print
statements in for now so that if the tests randomly fail again we
can get some more insight for further debugging.