Split large blocks to fit on one neuron core
This commit introduces a new build step that splits blocks
produced by the build process that would fail validation
because they cannot fit on a single neuron core. Each large
block is split into smaller blocks that can fit on a single
neuron core. This happens automatically when a model fundamentally
cannot be made to fit, which most commonly occurs when an
ensemble has more than 1024 neurons. However, it can also be occur
manually with the new `block_shape` Ensemble configuration option,
which is added in `add_params`. This parameter accepts a
`BlockShape` instance, which specifies how the ensemble should be
split. See the `BlockShape` documentation for more details.

Adding a new build step also added more complexity to the
Simulator.__init__ method, so we moved many of the build-related
steps that were previously in __init__ to the `build_network`
function. In doing so, it is now harder to half-build a model,
and the Simulator.__init__ method is less complex.

Note that the block splitting process occurs before discretization,
which means that this commit can change the behavior of existing
models by specifying the `block_shape` of an ensemble; large
models were not possible previously, so no existing models
will change because blocks are only automatically split if they
are too large to fit on one neuron core. We believe that doing
discretization after splitting will only yield improvements,
but we have not rigorously tested that belief.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>