Modify probes to work with multiple targets
This is required to allow probing across multiple blocks.
Since this means that probes are not specific to one block,
we should no longer consider them as part of a block, and
instead the mental model is that they lie alongside the
`LoihiInput` and `LoihiBlock` in the hierarchy of objects.
So, a Loihi Model is a collection of `LoihiBlock`s that get
input from `LoihiInput`s and get output from `LoihiProbe`s.
Accordingly, we've renamed the `Probe` to `LoihiProbe` for
consistency, and moved it a new top-level file, probe.py.

Co-authored-by: Kinjal Patel <kinjalp.patel@mail.utoronto.ca>

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>