Add host Snip for faster communication
The host Snip runs on the host and facilitates communication
with the superhost using sockets. This is faster than using
the default RPC interface.

We also take care to make sure both the host and chip Snips
end properly, by sending a message with a negative spike count.
This helps to eliminate board hangs.

To allow the host Snip to work with multiple `run` calls, we
keep it idling in between `run` calls, waiting for a message.
If the board disconnects before a subsequent run call,
the negative spike count message will tell the host Snip to stop.

This commit also includes several other improvements that either
are necessary for implementing the host snip, or are complementary
to it.

- The channel packet size in the IO Snip has been increased.
  This improves performance by reducing the number of channel reads.
- The logic for generating the correct run_steps method has been
  moved to a separate class for readability.
- The host2chip logic when using snips is now faster due to
  getting the core outside of the loop in the learn Snip.
- We now connect to the board when entering the simulator's context
  manager.
- Snip and non-snip operations have been separated into separate
  classes to make it more clear what is run in the two cases.
- Minor fixes to the style of C/C++ templates using clang-format
  with `-style="{BasedOnStyle: 'LLVM', IndentWidth: 4, TabWidth: 4}`.
  Since clang-format must be run on the generated files rather than
  the Jinja templates, this is difficult to automate, but is worth
  doing every once in a while.
- `nengo_io.c` has been removed from the gitignore, as the template
  should only ever be rendered to a temporary directory.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>