Add host SNIP for faster communication
The host SNIP runs on the host and facilitates communication
with the superhost using sockets. This is faster than using
the default RPC interface.

We also take care to make sure both the host and chip SNIPs
end properly, by sending a message with a negative spike count.
This helps to eliminate board hangs.

To allow the host SNIP to work with multiple `run` calls, we
keep it idling in between `run` calls, waiting for a message.
If the board disconnects before a subsequent run call,
the negative spike count message will tell the host SNIP to stop.