Snips work across multiple chips
This commit makes it possible to use multi-chip allocators in
non-precomputable models. Combined with recent commits that split
large blocks, it is now possible to run much larger real-world
models on Loihi.

Major changes
-------------

The superhost, host, and chip now all attempt to shut down when
one of the other devices shuts down. This is signaled by a -1
sent to/from the superhost<->host and host<->chip sockets.

Since race conditions can occur, the chip Snip waits at the end
of the simulation before signalling the end of the simulation.
This allows other chips to catch up and finish using the channels
before they close.

We also fix a bug in the error message when an unsupported neuron type
is requested.

Refactoring
-----------

The `Snips` class has been rewritten to support sending data to
and receiving data from multiple chips. Several new classes were
added to make the `Snips` class readable, including `ChipSnips`,
and `nengo_loihi.hardware.builder.ProbeSnip`. The `HostSocket`
class has been changed to a `HostSnip` class that encapsulates
the logic for both the Snip and the socket.

More Snip-specific logic has been moved from `HardwareInterface`
to `Snips`.

We pre-deobfuscate in build_synapse as deobfuscating within a loop
is expensive.

We remove the `Probe.use_snips` and `Probe.snip_info` attributes
because they contain hardware-specific information that was only
added in the hardware build process in order to be read in a later
part of the hardware build process. Instead, this information is
either stored in the model or passed through to the `build_probe`
function.

Shutdown behavior in Snips that used to be handled with jumps to
a `shutdown` label are now handled with function calls instead.

All host snip messages are now prepended with "[Host] " to make it
easier to scan debug logs. In order to make this possible, we switch
from using std::cout (<iostream>) to printf.

Co-authored-by: Eric Hunsberger <eric.hunsberger@appliedbrainresearch.com>

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>