Fix socket overloading
The socket between the superhost and host was dropping data when
trying to send larger numbers of spikes. This seems to be solved
by getting rid of the step counter on the host snip. Something about
sending the number of steps as a separate message at the start
threw things off in the socket, I guess.