Temporary workaround for multi-chip SNIP hanging
When using SNIPs with multiple chips, the simulation would hang on
the last timestep. This is because the first chip would finish before
the other chips, which would in turn close the channels.

To work around this, we wait at the end of the last timestep in the
first chip's SNIP. This allows the other chips to catch up and
finish using the channels before they close.