Report full board connection error
The reason for not connecting can be any error that NxSDK throws
when trying to connect, including problems with building the model,
as well as actual connectivity problems. We now include the actual
NxSDK error message in our error message to better inform users.

Also reduce the number of connection attempts, since if it fails
more than 3 times it's likely due to a model problem, not a
connectivity problem.