Support larger convolution transform outputs
We avoid the restriction of the axon compartment bases being less
than 256, by taking any part of the required offset over this
and adding it into the indices instead.

We also clarify the `conv2d_loihi_weights` function and
remove nengo-dl from `test_conv.py:test_conv_connection`.
The nengo-dl portions have not been run in a long time, since they
were only enabled if `channels_last=True`, and we were x-failing
the test for `channels_last=True` due to lack of nengo-loihi
support. It contributes nothing to testing nengo-loihi, so we remove it.