Simplify passthrough removal using HostChipSplit
Previously, we accepted an `ignore` parameter that had to be
manually constructed. However, the only time this information
ever really came into play was when an object was on chip.
By passing in a `HostChipSplit` object (which is easy to construct)
we get the same information automatically.

Additionally, this commit calls `config.add_params` in HostChipSplit,
rather than relying on it being called earlier. There is no downside
to calling `add_params` twice, so we just call it, removing the need
to call it in the Simulator and in several test functions.