Make n_chips a hardware option
Previously, n_chips was an argument to allocators. However, that
means that to tell Nengo Loihi that you want to use more chips,
you have to know that this is defined in the allocators, and then
you have to decide which allocator you want. Making it part of the
hardware options (i.e., HardwareInterface constructor) makes it
easier for users to specify, and reduces duplication because all
current (and likely future) allocators are multi-chip. The number
of chips is still available to allocators, as it is passed
into `__call__`.