Wait longer to get all chip's data in host snip
Previously, the host snip would wait a maximum of 10 milliseconds
to get data from all chips, regardless of the number of chips.
This commit relaxes it so that each chip has 100 milliseconds
to respond, and if one chip is responding that resets the clock
for the other chips. The 10 millisecond time limit was too
stringent for a few of the tests in the test suite.

This commit also improves how we clean up TravisCI builds
by using `rsync -a --delete` instead of `rm -r` and piping
the output to `tqdm` in order to track progress in the
console, which should stop builds from timing out due to
no console output happening.

This commit also adds several debugging log messages and print
statements in Snips that were helpful for debugging the test
failures that this commit will hopefully fix.